requirements.txt
python-dotenv>=1.0.1

# LLM Providers
openai>=1.3.0
anthropic>=0.7.0
google-generativeai>=0.3.0

# Search Providers
google-search-results>=2.4.2
duckduckgo-search>=3.9.0
aiohttp>=3.8.0
ddg